Extract ImagesTableRow into own component
